Refactor main.py into multiple modules
I've decomposed the lengthy main.py into several specialized modules to improve your code's organization, readability, and maintainability:

- `app_setup.py`: Contains global configurations, constants, application setup utilities (logging, stylesheet loading), and default template initialization logic.
- `ui_components.py`: Houses common UI classes `StatisticsWidget` and `StatusDelegate`.
- `document_manager_logic.py`: Encapsulates complex business logic and callback functions previously part of the `DocumentManager` class, such as client creation, loading, filtering, and management routines.
- `main_window.py`: Contains the `DocumentManager` class, which defines the main application window and its UI structure.
- `main.py`: Serves as the streamlined main application entry point, responsible for initializing the Qt application, loading configurations, and launching the main window.

Associated changes:
- I updated import statements across all affected files and related modules (`client_widget.py`, `utils.py`) to reflect the new structure.
- I also cleaned up unused imports and improved dependency management in some modules (e.g., `client_widget.py`).
